requirements.txt
pyserial==3.5
paho-mqtt==1.5.1